+++ flink-clients/src/test/java/org/apache/flink/client/program/PerJobMiniClusterFactoryTest.java
@@ -22,12 +22,11 @@ import org.apache.flink.api.common.JobExecutionResult;
 import org.apache.flink.api.common.JobStatus;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.core.execution.JobClient;
-import org.apache.flink.runtime.execution.Environment;
 import org.apache.flink.runtime.jobgraph.JobGraph;
 import org.apache.flink.runtime.jobgraph.JobGraphTestUtils;
 import org.apache.flink.runtime.jobgraph.JobVertex;
 import org.apache.flink.runtime.minicluster.MiniCluster;
-import org.apache.flink.runtime.testutils.CancelableInvokable;
+import org.apache.flink.runtime.testutils.WaitingCancelableInvokable;
 import org.apache.flink.util.TestLogger;
 
 import org.junit.After;
@@ -169,36 +168,8 @@ public class PerJobMiniClusterFactoryTest extends TestLogger {
 
     private static JobGraph getCancellableJobGraph() {
         JobVertex jobVertex = new JobVertex("jobVertex");
-        jobVertex.setInvokableClass(MyCancellableInvokable.class);
+        jobVertex.setInvokableClass(WaitingCancelableInvokable.class);
         jobVertex.setParallelism(1);
         return JobGraphTestUtils.streamingJobGraph(jobVertex);
-    }
-
-    /** Invokable which waits until it is cancelled. */
-    public static class MyCancellableInvokable extends CancelableInvokable {
-
-        private final Object lock = new Object();
-        private boolean running = true;
-
-        public MyCancellableInvokable(Environment environment) {
-            super(environment);
-        }
-
-        @Override
-        public void doInvoke() throws Exception {
-            synchronized (lock) {
-                while (running) {
-                    lock.wait();
-                }
-            }
-        }
-
-        @Override
-        public void cancel() {
-            synchronized (lock) {
-                running = false;
-                lock.notifyAll();
-            }
-        }
     }
 }
